pyproject.toml
[project]
name = "instagram-content-generator"
version = "0.1.0"
description = "Automated Instagram content generator with AI-powered analysis and caption generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "instagrapi>=2.0.0",  # Instagram private API
    "pillow>=10.0.0",     # Image processing
    "opencv-python>=4.8.0",  # Video processing
    "transformers>=4.30.0",  # AI models for content analysis
    "torch>=2.0.0",       # PyTorch for ML models
    "torchvision>=0.15.0",  # Computer vision models
    "requests>=2.31.0",   # HTTP requests
    "python-dotenv>=1.0.0",  # Environment variables
    "schedule>=1.2.0",    # Task scheduling
    "watchdog>=3.0.0",    # File system monitoring
    "loguru>=0.7.0",      # Advanced logging
    "pydantic>=2.0.0",    # Data validation
    "pydantic-settings>=2.0.0",  # Settings management
    "openai>=1.0.0",      # OpenAI API for caption generation
    "moviepy>=1.0.3",     # Video file handling
    "python-magic>=0.4.27",  # File type detection
    "psutil>=5.9.0",      # System monitoring
    "scikit-learn>=1.3.0",  # Machine learning utilities
    "numpy>=1.24.0",      # Numerical computing
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[project.scripts]
instagram-content-generator = "src.main:main"
